Changed to  detail report poMatch to float

+++ app.py
@@ -28,8 +28,8 @@ detail = pd.read_csv(detail_file)
 
 #### Merge()  VLOOKUP Equivalent
 
-remit['poMatch']  = remit['Invoice Number']
-detail['poMatch']  = detail['Purchase Order Number']
+remit['poMatch']  = remit['Invoice Number'].astype(float)
+detail['poMatch']  = detail['Purchase Order Number'].astype(float)
 remit_review = remit[['poMatch','Amount Paid($)','Store Number','Invoice Date']]
 detail_review = detail[['poMatch','Ship-To Customer Name','Invoice Number','Requested Delivery Date']]
